Added distinction between valid and computable
The app did not make any clear difference
between boards that were not valid
and not (reasonably) computable.

A board that was not computable could still be valid.
The app now displays a message
that the board is valid, but that more clues are needed

Also added style changes and lowered the minimum amount of clues